+++ application/config/autoload.php
@@ -40,6 +40,7 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 |
 */
 $autoload['packages'] = array();
+$autoload['packages'] = array(APPPATH.'third_party/restful');
 
 /*
 | -------------------------------------------------------------------
@@ -58,7 +59,7 @@ $autoload['packages'] = array();
 |
 |	$autoload['libraries'] = array('user_agent' => 'ua');
 */
-$autoload['libraries'] = array('parser', 'parsedown', 'session');
+$autoload['libraries'] = array('parser', 'parsedown', 'session','curl', 'format', 'rest');
 
 /*
 | -------------------------------------------------------------------