+++ application/models/Tasks.php
@@ -4,7 +4,9 @@ class Tasks extends XML_Model {
 
         public function __construct()
         {
-                parent::__construct(APPPATH . '../data/tasks.xml', 'id');
+                //parent::__construct(APPPATH . '../data/tasks.xml', 'id');
+				parent::__construct();
+				$this->load->library(['curl', 'format', 'rest']);
         }
 		
 		public function getCategorizedTasks()
@@ -40,6 +42,57 @@ class Tasks extends XML_Model {
 			['field' => 'group', 'label' => 'Task group', 'rules' => 'integer|less_than[5]'],
 		);
 		return $config;
+	}
+	
+	protected function load() 
+	{
+		
+		
+		// load our data from the REST backend
+        $this->rest->initialize(array('server' => REST_SERVER));
+        $this->rest->option(CURLOPT_PORT, REST_PORT);
+        $this->_data =  $this->rest->get('/job');
+		
+		$this->reindex();
+	}
+	
+	protected function store()
+	{
+		
+	}
+	
+	function get($key, $key2 = null) 
+	{
+        $this->rest->initialize(array('server' => REST_SERVER));
+        $this->rest->option(CURLOPT_PORT, REST_PORT);
+        return $this->rest->get('/job/' . $key);
+	}
+	
+	function delete($key, $key2 = null)
+	{
+		if (isset($this->_data[$key]))
+		{
+			unset($this->_data[$key]);
+			$this->store();
+		}
+	}
+	
+	function update($record)
+	{
+        $this->rest->initialize(array('server' => REST_SERVER));
+        $this->rest->option(CURLOPT_PORT, REST_PORT);
+        $key = $record->{$this->_keyfield};
+        $retrieved = $this->rest->put('/job/' . $key, $record);
+        $this->load(); // because the "database" might have changed
+	}
+	
+	function add($record)
+	{
+        $this->rest->initialize(array('server' => REST_SERVER));
+        $this->rest->option(CURLOPT_PORT, REST_PORT);
+        $key = $record->{$this->_keyfield};
+        $retrieved = $this->rest->post('/job/' . $key, $record);
+        $this->load(); // because the "database" might have changed
 	}
 }